feat : add template creation implementation
